Refactor license module imports and paths to use plural 'licenses' instead of 'license'. This change affects multiple files across the API and proprietary directories, ensuring consistency in module references and improving clarity in the codebase.
