+++ proprietary/data-retention/__tests__/data-retention.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { DataRetentionService } from '../data-retention.service';
-import { LicenseService } from '@proprietary/license/license.service';
-import { Tier } from '@proprietary/license/types';
+import { LicenseService } from '@proprietary/licenses/license.service';
+import { Tier } from '@proprietary/licenses/types';
 import { StoragePort } from '@app/common/interfaces/storage-port.interface';
 
 const MS_PER_DAY = 24 * 60 * 60 * 1000;
