+++ proprietary/licenses/__tests__/license.service.spec.ts
