+++ proprietary/licenses/index.ts
