+++ proprietary/licenses/license.guard.ts
