+++ proprietary/licenses/license.module.ts
