+++ proprietary/licenses/license.service.spec.ts
