+++ proprietary/licenses/license.service.ts
