+++ proprietary/licenses/requires-feature.decorator.ts
