+++ proprietary/licenses/types.ts
