+++ proprietary/webhook-pro/webhook-events-pro.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, Logger, OnModuleInit, Inject } from '@nestjs/common';
 import { WebhookDispatcherService } from '@app/webhooks/webhook-dispatcher.service';
 import { WebhookEventType } from '@betterdb/shared';
-import { LicenseService } from '@proprietary/license';
+import { LicenseService } from '@proprietary/licenses';
 
 /**
  * Webhook Events Pro Service - Generates PRO tier webhook events
